+++ src/interfaces/ISentinel.cairo
@@ -7,8 +7,8 @@ pub trait ISentinel<TContractState> {
     fn get_gate_address(self: @TContractState, yang: ContractAddress) -> ContractAddress;
     fn get_gate_live(self: @TContractState, yang: ContractAddress) -> bool;
     fn get_yang_addresses(self: @TContractState) -> Span<ContractAddress>;
-    fn get_yang_addresses_count(self: @TContractState) -> u64;
-    fn get_yang(self: @TContractState, idx: u64) -> ContractAddress;
+    fn get_yang_addresses_count(self: @TContractState) -> u32;
+    fn get_yang(self: @TContractState, idx: u32) -> ContractAddress;
     fn get_yang_asset_max(self: @TContractState, yang: ContractAddress) -> u128;
     fn get_asset_amt_per_yang(self: @TContractState, yang: ContractAddress) -> Wad;
     // external
